+++ src/pages/Landing/components/MeetingCard.jsx
